fix(minifront): #1211: make the code cleaner after the review
