+++ apps/minifront/src/state/ibc-out.ts
@@ -14,8 +14,7 @@ import {
 } from '@penumbra-zone/getters/value-view';
 import { getAddressIndex } from '@penumbra-zone/getters/address-view';
 import { toBaseUnit } from '@penumbra-zone/types/lo-hi';
-import { planBuildBroadcast } from './helpers';
-import { amountMoreThanBalance } from './send';
+import { amountMoreThanBalance, planBuildBroadcast } from './helpers';
 import { getAssetId } from '@penumbra-zone/getters/metadata';
 import { assetPatterns } from '@penumbra-zone/types/assets';
 import { bech32, bech32m } from 'bech32';
