+++ apps/minifront/src/state/send.ts
@@ -6,7 +6,7 @@ import {
 } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/view/v1/view_pb';
 import { BigNumber } from 'bignumber.js';
 import { MemoPlaintext } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/transaction/v1/transaction_pb';
-import { plan, planBuildBroadcast } from './helpers';
+import { amountMoreThanBalance, plan, planBuildBroadcast } from './helpers';
 
 import {
   Fee,
@@ -18,7 +18,6 @@ import {
 } from '@penumbra-zone/getters/value-view';
 import { getAddress, getAddressIndex } from '@penumbra-zone/getters/address-view';
 import { toBaseUnit } from '@penumbra-zone/types/lo-hi';
-import { fromValueView } from '@penumbra-zone/types/amount';
 import { isAddress } from '@penumbra-zone/bech32m/penumbra';
 import { ZQueryState, createZQuery } from '@penumbra-zone/zquery';
 import { getTransferableBalancesResponses } from '../components/send/helpers';
@@ -169,22 +168,6 @@ const assembleRequest = ({ amount, feeTier, recipient, selection, memo }: SendSl
       text: memo,
     }),
   });
-};
-
-export const amountMoreThanBalance = (
-  asset: BalancesResponse,
-  /**
-   * The amount that a user types into the interface will always be in the
-   * display denomination -- e.g., in `penumbra`, not in `upenumbra`.
-   */
-  amountInDisplayDenom: string,
-): boolean => {
-  if (!asset.balanceView) {
-    throw new Error('Missing balanceView');
-  }
-
-  const balanceAmt = fromValueView(asset.balanceView);
-  return Boolean(amountInDisplayDenom) && BigNumber(amountInDisplayDenom).gt(balanceAmt);
 };
 
 export interface SendValidationFields {
