+++ apps/minifront/src/state/swap/index.ts
@@ -15,7 +15,6 @@ import {
   createInstantSwapSlice,
   InstantSwapSlice,
   instantSwapSubmitButtonDisabledSelector,
-  isValidAmount,
 } from './instant-swap';
 import { createPriceHistorySlice, PriceHistorySlice } from './price-history';
 import { getMetadata } from '@penumbra-zone/getters/value-view';
@@ -23,6 +22,7 @@ import { createZQuery, ZQueryState } from '@penumbra-zone/zquery';
 import { getSwappableBalancesResponses, isSwappable } from '../../components/swap/helpers';
 import { getAllAssets } from '../../fetchers/assets';
 import { emptyBalanceResponse } from '../../utils/empty-balance-response';
+import { isValidAmount } from '../helpers';
 
 // When both `balancesResponses` and `swappableAssets` are loaded, set initial assetIn and assetOut
 const setInitialAssets = (state: SwapSlice) => {
